feat: add form validation
