+++ messages/uk/components/ContactForm.json
@@ -6,5 +6,8 @@
   "send": "Відправити повідомлення",
   "sending": "Відправка...",
   "success": "Повідомлення надіслано!",
-  "error": "Не вдалося надіслати повідомлення"
+  "error": "Не вдалося надіслати повідомлення",
+  "errorValidateForm": "Будь ласка, заповніть форму",
+  "errorValidateField": "Це поле є обов’язковим.",
+  "errorValidateFiledEmail": "Некоректна електронна адреса"
 }
